Database tests base context
