+++ Backtrace.Tests/DatabaseTests/Model/DatabaseTestBase.cs
@@ -1,0 +1,70 @@
+﻿using Backtrace.Base;
+using Backtrace.Interfaces;
+using Backtrace.Model;
+using Backtrace.Model.Database;
+using Backtrace.Types;
+using Moq;
+using NUnit.Framework;
+using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Backtrace.Tests.DatabaseTests.Model
+{
+    /// <summary>
+    /// Database tests base class
+    /// </summary>
+    public class DatabaseTestBase
+    {
+        /// <summary>
+        /// Database
+        /// </summary>
+        protected BacktraceDatabase<object> _database;
+
+        [SetUp]
+        public void Setup()
+        {
+            //get project path
+            string projectPath = Environment.CurrentDirectory;
+
+            //mock api
+            var mockApi = new Mock<IBacktraceApi<object>>();
+            mockApi.Setup(n => n.Send(It.IsAny<BacktraceData<object>>()))
+                .Returns(new BacktraceResult());
+
+            //mock file context
+            var mockFileContext = new Mock<IBacktraceDatabaseFileContext<object>>();
+            mockFileContext.Setup(n => n.GetEntries())
+                .Returns(new List<FileInfo>());
+
+            //mock cache
+            var mockCacheContext = new Mock<IBacktraceDatabaseContext<object>>();
+            mockFileContext.Setup(n => n.RemoveOrphaned(It.IsAny<IEnumerable<BacktraceDatabaseEntry<object>>>()));
+
+            _database = new BacktraceDatabase<object>(projectPath)
+            {
+                BacktraceDatabaseContext = new MockBacktraceDatabaseContext(projectPath, 3, RetryOrder.Stack),
+                BacktraceDatabaseFileContext = mockFileContext.Object,
+                BacktraceApi = mockApi.Object
+            };
+        }
+
+        /// <summary>
+        /// Get new database entry 
+        /// </summary>
+        /// <returns>Database entry mock</returns>
+        protected BacktraceDatabaseEntry<object> GetEntry()
+        {
+            //mock single entry
+            var mockEntry = new Mock<BacktraceDatabaseEntry<object>>();
+            mockEntry.Setup(n => n.Delete());
+            mockEntry.Setup(n => n.BacktraceData)
+                .Returns(new BacktraceData<object>(It.IsAny<BacktraceReportBase<object>>(), It.IsAny<Dictionary<string, object>>()));
+
+            return mockEntry.Object;
+        }
+    }
+}
